Настройка уведомлений и еще немного работы с бд

+++ database.py
@@ -1,3 +1,4 @@
+import datetime
 import sqlite3
 
 connection = sqlite3.connect('groups.db')
@@ -11,53 +12,60 @@ groups TEXT NOT NULL) ''')
 cursor.execute('''
 CREATE TABLE IF NOT EXISTS Weekly (
 id INTEGER PRIMARY KEY,
-time TIME) ''')
+time TIME NOT NULL,
+enabled BIT NOT NULL) ''')
 
 cursor.execute('''
 CREATE TABLE IF NOT EXISTS Tomorrow (
 id INTEGER PRIMARY KEY,
-time TIME) ''')
+time TIME NOT NULL,
+enabled BIT NOT NULL) ''')
 
 cursor.execute('''
 CREATE TABLE IF NOT EXISTS Today (
 id INTEGER PRIMARY KEY,
-time TIME) ''')
+time TIME NOT NULL,
+enabled BIT NOT NULL) ''')
 
-def check_source(source) -> True | ValueError:
+def check_source(source: str ,include_groups: bool = True) -> True:
 	"""
 	Checks the spelling of arguments for functions / Проверяет правильность написания аргументов функций
+	:param include_groups: include 'Groups' to allowed list of values / Добавить 'Groups' в список разрешенных значений
 	:param source: table of DB that need to check spell / Таблица БД написание которой надо проверить
 	:return: True if spell is right else ValueError / True если все верно, иначе ValueError
 	"""
-	if source not in ("Groups","Weekly","Tomorrow","Today"):
+	source_filter=("Groups","Weekly","Tomorrow","Today") if include_groups else ("Weekly","Tomorrow","Today")
+	if source not in source_filter:
 		raise ValueError(f"Некорректное название таблицы {source} для поиска")
 	return True
 
-def add_to(source, chat_id, time) -> None:
+def add_to(source: str , chat_id: int, time: str) -> None:
 	"""
-	Adds chat id to DB to use timetable notifications /Добавляет ID чата в телеграмме для дальнейшего использования
+	Adds chat id to DB to use schedule notifications /Добавляет ID чата в телеграмме для дальнейшего использования
 	:param source: table of DB where it's need to search /Таблица БД где нужно искать
-	:param chat_id: ID of telegram chat where bot will send timetable /ID чата в телеграмме где бот будет публиковать расписание
-	:param time: time when it's need to send timetable / Время когда надо присылать расписание
+	:param chat_id: ID of telegram chat where bot will send schedule /ID чата в телеграмме где бот будет публиковать расписание
+	:param time: time when it's need to send schedule / Время когда надо присылать расписание
 	:return: None
 	"""
-	check_source(source)
-	cursor.execute(f'INSERT INTO {source} (id, time) VALUES (?, ?)', (chat_id,time))
+	check_source(source,include_groups=False)
+	cursor.execute(f'INSERT INTO {source} (id, time, enabled) VALUES (?, ?, 1)', (chat_id,time))
 
-def set_group(chat_id,group) -> None | int:
+def set_group(chat_id: int , group: str) -> None:
 	"""
-	Adds chat id to DB to use timetable notifications /Добавляет ID чата в телеграмме для дальнейшего использования
-	:param chat_id: ID of telegram chat where bot will send timetable /ID чата в телеграмме где бот будет публиковать расписание
+	Adds chat id to DB to use schedule notifications /Добавляет ID чата в телеграмме для дальнейшего использования
+	:param chat_id: ID of telegram chat where bot will send schedule /ID чата в телеграмме где бот будет публиковать расписание
 	:param group: group of college /Группа колледжа, для которой нужно искать расписание
 	:return: None
 	"""
-	cursor.execute('INSERT INTO Groups (id,groups) VALUES (?, ?)'
-	               'ON CONFLICT(id) DO UPDATE SET groups = excluded.groups', (chat_id,group))
-	for source in ("Weekly","Tomorrow","Today"):
-		add_to(source,chat_id,"08:00")
+	if not is_in("Groups",chat_id):
+		for source in ("Weekly", "Tomorrow", "Today"):
+			add_to(source, chat_id, "08:00:00")
+		cursor.execute('INSERT INTO Groups (id,groups) VALUES (?, ?)', (chat_id, group))
+	else:
+		cursor.execute('UPDATE Groups SET groups = ? WHERE id = ?', (group,chat_id))
 	connection.commit()
 
-def is_in(chat_id,source) -> bool:
+def is_in(source: str , chat_id: int) -> bool:
 	"""
 	Returns is chat ID in table of DB / Возвращает, есть ли ID чата в таблице БД
 	:param chat_id: ID of telegram chat / ID чата в телеграмм
@@ -68,7 +76,7 @@ def is_in(chat_id,source) -> bool:
 	cursor.execute(f"SELECT * FROM {source} WHERE id=?", (chat_id,))
 	return bool(cursor.fetchone())
 
-def return_from(chat_id, source) -> str:
+def return_from(source: str,chat_id: int) -> dict:
 	"""
 	Returns time/group in table of DB / Возвращает группу/время из таблицы БД
 	:param chat_id: ID of telegram chat / ID чата в телеграмм
@@ -77,9 +85,9 @@ def return_from(chat_id, source) -> str:
 	"""
 	check_source(source)
 	cursor.execute(f"SELECT * FROM {source} WHERE id=?", (chat_id,))
-	return cursor.fetchall()[0]
+	return dict(zip(("id","time","enabled"),cursor.fetchone()))
 
-def return_table(source) -> list:
+def return_table(source: str) -> list:
 	"""
 	Returns table of DB / Возвращает таблицу БД
 	:param source: table of DB where it's need to return / Таблица БД которую нужно вернуть
@@ -89,14 +97,31 @@ def return_table(source) -> list:
 	cursor.execute(f"SELECT * FROM {source}")
 	return cursor.fetchall()
 
-def remove_from(chat_id,source) -> None:
+def remove_from(source: str, chat_id: int) -> None:
 	"""
 	Deletes data in table of DB/ Удаляет данные в таблице БД
 	:param chat_id: ID of telegram chat / ID чата в телеграмм
 	:param source: table of DB where it's need to search / Таблица БД где нужно искать
 	:return: None
 	"""
+	if not is_in(source,chat_id):
+		raise ValueError(f"Указанного ID '{chat_id}' нет в таблице БД")
 	check_source(source)
 	cursor.execute(f'DELETE FROM {source} WHERE id = ?', (chat_id,))
 	connection.commit()
 
+def is_enabled(source: str, chat_id: int) -> bool:
+	if not is_in(source,chat_id):
+		raise ValueError(f"Указанного ID '{chat_id}' нет в таблице БД")
+	check_source(source, include_groups=False)
+	return bool(return_from(source,chat_id)["enabled"])
+
+def turn_notification(source: str, chat_id: int) -> None:
+	enabled=int(not(is_enabled(source,chat_id)))
+	cursor.execute(f'UPDATE {source} SET enabled = ? WHERE id = ?', (enabled,chat_id))
+	connection.commit()
+
+def change_time(source: str, chat_id: int, time: datetime.time) -> None:
+	check_source(source,include_groups=False)
+	cursor.execute(f'UPDATE {source} SET time = ? WHERE id = ?', (str(time),chat_id))
+	connection.commit()
